pyproject.toml
[project]
name = "fastapi-otel-loki-grafana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "python-json-logger>=3.3.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
fastapi-server = "app.__main__:main"